Feature/weos 1132 - As a developer I should be able to update an item for a content type

+++ end2end_test.go
@@ -7,6 +7,15 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"net/http"
+	"net/http/httptest"
+	"os"
+	"path/filepath"
+	"strconv"
+	"strings"
+	"testing"
+	"time"
+
 	"github.com/cucumber/godog"
 	"github.com/labstack/echo/v4"
 	ds "github.com/ompluscator/dynamic-struct"
@@ -15,13 +24,6 @@ import (
 	api "github.com/wepala/weos/controllers/rest"
 	"github.com/wepala/weos/utils"
 	"gorm.io/gorm"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"path/filepath"
-	"strings"
-	"testing"
-	"time"
 )
 
 var e *echo.Echo
@@ -33,6 +35,7 @@ var errs error
 var buf bytes.Buffer
 var payload ContentType
 var rec *httptest.ResponseRecorder
+var header http.Header
 var resp *http.Response
 var db *sql.DB
 var requests map[string]map[string]interface{}
@@ -68,6 +71,7 @@ type ContentType struct {
 
 func InitializeSuite(ctx *godog.TestSuiteContext) {
 	requests = map[string]map[string]interface{}{}
+	contentTypeID = map[string]bool{}
 	Developer = &User{}
 	e = echo.New()
 	e.Logger.SetOutput(&buf)
@@ -116,8 +120,10 @@ components:
 
 func reset(ctx context.Context, sc *godog.Scenario) (context.Context, error) {
 	requests = map[string]map[string]interface{}{}
+	contentTypeID = map[string]bool{}
 	Developer = &User{}
 	errs = nil
+	header = make(http.Header)
 	rec = httptest.NewRecorder()
 	resp = nil
 	os.Remove("e2e.db")
@@ -316,6 +322,53 @@ func anErrorShouldBeReturned() error {
 }
 
 func blogsInTheApi(details *godog.Table) error {
+
+	head := details.Rows[0].Cells
+
+	for i := 1; i < len(details.Rows); i++ {
+		req := make(map[string]interface{})
+		seq := 0
+		for n, cell := range details.Rows[i].Cells {
+			if (head[n].Value) != "sequence_no" {
+				req[head[n].Value] = cell.Value
+			} else {
+				seq, _ = strconv.Atoi(cell.Value)
+			}
+		}
+		reqBytes, _ := json.Marshal(req)
+		body := bytes.NewReader(reqBytes)
+		var request *http.Request
+
+		request = httptest.NewRequest("POST", "/blog", body)
+
+		request = request.WithContext(context.TODO())
+		header.Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
+		request.Header = header
+		request.Close = true
+		rec = httptest.NewRecorder()
+		e.ServeHTTP(rec, request)
+		if rec.Code != http.StatusCreated {
+			return fmt.Errorf("expected the status to be %d got %d", http.StatusCreated, rec.Code)
+		}
+
+		if seq > 1 {
+			reqBytes, _ := json.Marshal(req)
+			body := bytes.NewReader(reqBytes)
+			for i := 1; i < seq; i++ {
+				request = httptest.NewRequest("PUT", "/blogs/"+req["id"].(string), body)
+				request = request.WithContext(context.TODO())
+				header.Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
+				request.Header = header
+				request.Close = true
+				rec = httptest.NewRecorder()
+				e.ServeHTTP(rec, request)
+				if rec.Code != http.StatusOK {
+					return fmt.Errorf("expected the status to be %d got %d", http.StatusOK, rec.Code)
+				}
+			}
+		}
+
+	}
 	return nil
 }
 
@@ -356,8 +409,14 @@ func theIsCreated(contentType string, details *godog.Table) error {
 	}
 
 	contentEntity := map[string]interface{}{}
-	idEtag, seqNoEtag := api.SplitEtag(rec.Result().Header.Get("Etag"))
-	result := API.Application.DB().Table(strings.Title(contentType)).Find(&contentEntity, "weos_id = ?", idEtag, "sequence_no = ?", seqNoEtag)
+	var result *gorm.DB
+	//ETag would help with this
+	for key, value := range compare {
+		result = API.Application.DB().Table(strings.Title(contentType)).Find(&contentEntity, key+" = ?", value)
+		if contentEntity != nil {
+			break
+		}
+	}
 
 	if contentEntity == nil {
 		return fmt.Errorf("unexpected error finding content type in db")
@@ -373,6 +432,7 @@ func theIsCreated(contentType string, details *godog.Table) error {
 		}
 	}
 
+	contentTypeID[strings.ToLower(contentType)] = true
 	return nil
 }
 
@@ -385,9 +445,15 @@ func theIsSubmitted(contentType string) error {
 
 	reqBytes, _ := json.Marshal(req)
 	body := bytes.NewReader(reqBytes)
-	request := httptest.NewRequest("POST", "/"+strings.ToLower(contentType), body)
+	var request *http.Request
+	if strings.Contains(currScreen, "create") {
+		request = httptest.NewRequest("POST", "/"+strings.ToLower(contentType), body)
+	} else if strings.Contains(currScreen, "update") {
+		request = httptest.NewRequest("PUT", "/"+strings.ToLower(contentType)+"s/"+fmt.Sprint(req["id"]), body)
+	}
 	request = request.WithContext(context.TODO())
-	request.Header.Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
+	header.Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
+	request.Header = header
 	request.Close = true
 	rec = httptest.NewRecorder()
 	e.ServeHTTP(rec, request)
@@ -396,11 +462,9 @@ func theIsSubmitted(contentType string) error {
 
 func theShouldHaveAnId(contentType string) error {
 
-	idEtag, _ := api.SplitEtag(rec.Result().Header.Get("Etag"))
-	if idEtag == "" {
-		return fmt.Errorf("expected the "+contentType+" to have an ID, got %s", idEtag)
+	if !contentTypeID[strings.ToLower(contentType)] {
+		return fmt.Errorf("expected the " + contentType + " to have an ID")
 	}
-
 	return nil
 }
 
@@ -449,16 +513,16 @@ func aEntityConfigurationShouldBeSetup(arg1 string, arg2 *godog.DocString) error
 		field := reader.GetField(strings.Title(fields[1]))
 		switch fields[0] {
 		case "string":
-			if field.Interface() != "" {
+			if field.Interface() != "" && field.Interface() != field.PointerString() {
 				return fmt.Errorf("expected a string, got '%v'", field.Interface())
 			}
 
 		case "integer":
-			if field.Interface() != 0 {
+			if field.Interface() != 0 && field.Interface() != field.PointerInt() {
 				return fmt.Errorf("expected an integer, got '%v'", field.Interface())
 			}
 		case "uint":
-			if field.Interface() != uint(0) {
+			if field.Interface() != uint(0) && field.Interface() != field.PointerUint() {
 				return fmt.Errorf("expected an uint, got '%v'", field.Interface())
 			}
 		case "datetime":
@@ -474,6 +538,11 @@ func aEntityConfigurationShouldBeSetup(arg1 string, arg2 *godog.DocString) error
 
 	}
 
+	return nil
+}
+
+func aHeaderWithValue(key, value string) error {
+	header.Add(key, value)
 	return nil
 }
 
@@ -543,15 +612,25 @@ func isRunOnTheOperatingSystemAs(arg1 string, arg2 string) error {
 }
 
 func theEndpointIsHit(method, contentType string) error {
-	reqBytes, _ := json.Marshal(reqBody)
-	body := bytes.NewReader(reqBytes)
-	request := httptest.NewRequest(method, dockerEndpoint+contentType, body)
-	request = request.WithContext(context.TODO())
-	request.Header.Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
-	request.Close = true
-	client := http.Client{}
-	resp, errs = client.Do(request)
-	defer esContainer.Terminate(context.Background())
+	if binary != "" {
+		reqBytes, _ := json.Marshal(reqBody)
+		body := bytes.NewReader(reqBytes)
+		request := httptest.NewRequest(method, dockerEndpoint+contentType, body)
+		request = request.WithContext(context.TODO())
+		request.Header.Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
+		request.Close = true
+		client := http.Client{}
+		resp, errs = client.Do(request)
+		defer esContainer.Terminate(context.Background())
+	} else {
+		request := httptest.NewRequest(method, contentType, nil)
+		request = request.WithContext(context.TODO())
+		header.Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
+		request.Header = header
+		request.Close = true
+		rec = httptest.NewRecorder()
+		e.ServeHTTP(rec, request)
+	}
 	return nil
 }
 
@@ -568,19 +647,125 @@ func theServiceIsRunning() error {
 	return nil
 }
 
-func theHeaderShouldBe(header, value string) error {
-	//Table has no value to compare against so just checking for id existance
-	Etag := rec.Result().Header.Get(header)
-	idEtag, seqNoEtag := api.SplitEtag(Etag)
-	if Etag == "" {
-		return fmt.Errorf("expected the Etag to be added to header, got %s", Etag)
+func isOnTheEditScreenWithId(user, contentType, id string) error {
+	requests[strings.ToLower(contentType+"_update")] = map[string]interface{}{}
+	currScreen = strings.ToLower(contentType + "_update")
+	requests[currScreen]["id"] = id
+	return nil
+}
+
+func theHeaderShouldBe(key, value string) error {
+	if key == "ETag" {
+		Etag := rec.Result().Header.Get(key)
+		idEtag, seqNoEtag := api.SplitEtag(Etag)
+		if Etag == "" {
+			return fmt.Errorf("expected the Etag to be added to header, got %s", Etag)
+		}
+		if idEtag == "" {
+			return fmt.Errorf("expected the Etag to contain a weos id, got %s", idEtag)
+		}
+		if seqNoEtag == "" {
+			return fmt.Errorf("expected the Etag to contain a sequence no, got %s", seqNoEtag)
+		}
+
+		if seqNoEtag != strings.Split(value, ".")[1] {
+			return fmt.Errorf("expected the Etag to contain a sequence no %s, got %s", strings.Split(value, ".")[1], seqNoEtag)
+		}
+		return nil
 	}
-	if idEtag == "" {
-		return fmt.Errorf("expected the Etag to contain a weos id, got %s", idEtag)
+
+	headers := rec.Result().Header
+	val := []string{}
+
+	for k, v := range headers {
+		if strings.EqualFold(k, key) {
+			val = v
+			break
+		}
 	}
-	if seqNoEtag == "" {
-		return fmt.Errorf("expected the Etag to contain a sequence no, got %s", seqNoEtag)
+
+	if len(val) > 0 {
+		if strings.EqualFold(val[0], value) {
+			return nil
+		}
 	}
+	return fmt.Errorf("expected the header %s value to be %s got %v", key, value, val)
+}
+
+func theIsUpdated(contentType string, details *godog.Table) error {
+	if rec.Result().StatusCode != http.StatusOK {
+		return fmt.Errorf("expected the status code to be '%d', got '%d'", http.StatusOK, rec.Result().StatusCode)
+	}
+
+	head := details.Rows[0].Cells
+	compare := map[string]interface{}{}
+
+	for i := 1; i < len(details.Rows); i++ {
+		for n, cell := range details.Rows[i].Cells {
+			compare[head[n].Value] = cell.Value
+		}
+	}
+
+	contentEntity := map[string]interface{}{}
+	var result *gorm.DB
+	//ETag would help with this
+	for key, value := range compare {
+		result = API.Application.DB().Table(strings.Title(contentType)).Find(&contentEntity, key+" = ?", value)
+		if contentEntity != nil {
+			break
+		}
+	}
+	if contentEntity == nil {
+		result = API.Application.DB().Table(strings.Title(contentType)).Find(&contentEntity, "id = ?", requests[currScreen])
+	}
+
+	if contentEntity == nil {
+		return fmt.Errorf("unexpected error finding content type in db")
+	}
+
+	if result.Error != nil {
+		return fmt.Errorf("unexpected error finding content type: %s", result.Error)
+	}
+
+	for key, value := range compare {
+		if contentEntity[key] != value {
+			return fmt.Errorf("expected %s %s %s, got %s", contentType, key, value, contentEntity[key])
+		}
+	}
+
+	contentTypeID[strings.ToLower(contentType)] = true
+	return nil
+}
+
+func aBlogShouldBeReturned(details *godog.Table) error {
+	head := details.Rows[0].Cells
+	compare := map[string]interface{}{}
+
+	for i := 1; i < len(details.Rows); i++ {
+		for n, cell := range details.Rows[i].Cells {
+			compare[head[n].Value] = cell.Value
+		}
+	}
+
+	contentEntity := map[string]interface{}{}
+	err := json.NewDecoder(rec.Body).Decode(&contentEntity)
+
+	if err != nil {
+		return err
+	}
+
+	for key, value := range compare {
+		if contentEntity[key] != value {
+			return fmt.Errorf("expected %s %s %s, got %s", "Blog", key, value, contentEntity[key])
+		}
+	}
+
+	return nil
+}
+
+func sojournerIsUpdatingWithId(contentType, id string) error {
+	requests[strings.ToLower(contentType+"_update")] = map[string]interface{}{"id": id}
+	currScreen = strings.ToLower(contentType + "_update")
 	return nil
 }
 
@@ -607,9 +792,16 @@ func InitializeScenario(ctx *godog.ScenarioContext) {
 	ctx.Step(`^the "([^"]*)" should have an id$`, theShouldHaveAnId)
 	ctx.Step(`^the specification is$`, theSpecificationIs)
 	ctx.Step(`^the "([^"]*)" specification is parsed$`, theSpecificationIsParsed)
-	ctx.Step(`^a "([^"]*)" entity configuration should be setup$`, aEntityConfigurationShouldBeSetup)
-	ctx.Step(`^a warning should be output to logs letting the developer know that a parameter for each part of the idenfier must be set$`, aWarningShouldBeOutputToLogsLettingTheDeveloperKnowThatAParameterForEachPartOfTheIdenfierMustBeSet)
 	ctx.Step(`^the "([^"]*)" header should be "([^"]*)"$`, theHeaderShouldBe)
+	ctx.Step(`^a "([^"]*)" entity configuration should be setup$`, aEntityConfigurationShouldBeSetup)
+	ctx.Step(`^"([^"]*)" is on the "([^"]*)" edit screen with id "([^"]*)"$`, isOnTheEditScreenWithId)
+	ctx.Step(`^the "([^"]*)" is updated$`, theIsUpdated)
+	ctx.Step(`^a header "([^"]*)" with value "([^"]*)"$`, aHeaderWithValue)
+	ctx.Step(`^a (\d+) response should be returned$`, aResponseShouldBeReturned)
+	ctx.Step(`^the "([^"]*)" endpoint "([^"]*)" is hit$`, theEndpointIsHit)
+	ctx.Step(`^a blog should be returned$`, aBlogShouldBeReturned)
+	ctx.Step(`^Sojourner is updating "([^"]*)" with id "([^"]*)"$`, sojournerIsUpdatingWithId)
+	ctx.Step(`^a warning should be output to logs letting the developer know that a parameter for each part of the idenfier must be set$`, aWarningShouldBeOutputToLogsLettingTheDeveloperKnowThatAParameterForEachPartOfTheIdenfierMustBeSet)
 	ctx.Step(`^a "([^"]*)" route should be added to the api$`, aRouteShouldBeAddedToTheApi1)
 	ctx.Step(`^a (\d+) response should be returned$`, aResponseShouldBeReturned)
 	ctx.Step(`^request body$`, requestBody)
